+++ Gem/Code/Source/RHI/AsyncComputeExampleComponent.cpp
@@ -814,9 +814,9 @@ namespace AtomSampleViewer
                 drawItem.m_arguments = drawIndexed;
                 drawItem.m_pipelineState = m_copyTexturePipelineState.get();
                 drawItem.m_indexBufferView = &m_quadIndexBufferView;
-                drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+                drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
                 drawItem.m_shaderResourceGroups = shaderResourceGroups;
-                drawItem.m_streamBufferViewCount = static_cast<uint32_t>(m_quadStreamBufferViews[CopyTextureScope].size());
+                drawItem.m_streamBufferViewCount = static_cast<uint8_t>(m_quadStreamBufferViews[CopyTextureScope].size());
                 drawItem.m_streamBufferViews = m_quadStreamBufferViews[CopyTextureScope].data();
                 commandList->Submit(drawItem);
             }
@@ -879,9 +879,9 @@ namespace AtomSampleViewer
                 drawItem.m_arguments = drawIndexed;
                 drawItem.m_pipelineState = m_terrainPipelineStates[ShadowScope].get();
                 drawItem.m_indexBufferView = &m_quadIndexBufferView;
-                drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+                drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
                 drawItem.m_shaderResourceGroups = shaderResourceGroups;
-                drawItem.m_streamBufferViewCount = static_cast<uint32_t>(m_quadStreamBufferViews[ShadowScope].size());
+                drawItem.m_streamBufferViewCount = static_cast<uint8_t>(m_quadStreamBufferViews[ShadowScope].size());
                 drawItem.m_streamBufferViews = m_quadStreamBufferViews[ShadowScope].data();
                 commandList->Submit(drawItem);
             }
@@ -896,9 +896,9 @@ namespace AtomSampleViewer
                     drawItem.m_arguments = mesh.m_drawArguments;
                     drawItem.m_pipelineState = m_modelPipelineStates[ShadowScope].get();
                     drawItem.m_indexBufferView = &mesh.m_indexBufferView;
-                    drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+                    drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
                     drawItem.m_shaderResourceGroups = shaderResourceGroups;
-                    drawItem.m_streamBufferViewCount = static_cast<uint32_t>(m_modelStreamBufferViews[ShadowScope].size());
+                    drawItem.m_streamBufferViewCount = static_cast<uint8_t>(m_modelStreamBufferViews[ShadowScope].size());
                     drawItem.m_streamBufferViews = m_modelStreamBufferViews[ShadowScope].data();
 
                     commandList->Submit(drawItem);
@@ -994,9 +994,9 @@ namespace AtomSampleViewer
                 drawItem.m_arguments = drawIndexed;
                 drawItem.m_pipelineState = m_terrainPipelineStates[ForwardScope].get();
                 drawItem.m_indexBufferView = &m_quadIndexBufferView;
-                drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+                drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
                 drawItem.m_shaderResourceGroups = shaderResourceGroups;
-                drawItem.m_streamBufferViewCount = static_cast<uint32_t>(m_quadStreamBufferViews[ForwardScope].size());
+                drawItem.m_streamBufferViewCount = static_cast<uint8_t>(m_quadStreamBufferViews[ForwardScope].size());
                 drawItem.m_streamBufferViews = m_quadStreamBufferViews[ForwardScope].data();
 
                 commandList->Submit(drawItem);
@@ -1017,9 +1017,9 @@ namespace AtomSampleViewer
                     drawItem.m_arguments = mesh.m_drawArguments;
                     drawItem.m_pipelineState = m_modelPipelineStates[ForwardScope].get();
                     drawItem.m_indexBufferView = &mesh.m_indexBufferView;
-                    drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+                    drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
                     drawItem.m_shaderResourceGroups = shaderResourceGroups;
-                    drawItem.m_streamBufferViewCount = static_cast<uint32_t>(m_modelStreamBufferViews[ForwardScope].size());
+                    drawItem.m_streamBufferViewCount = static_cast<uint8_t>(m_modelStreamBufferViews[ForwardScope].size());
                     drawItem.m_streamBufferViews = m_modelStreamBufferViews[ForwardScope].data();
 
                     commandList->Submit(drawItem);
@@ -1167,9 +1167,9 @@ namespace AtomSampleViewer
                 drawItem.m_arguments = drawIndexed;
                 drawItem.m_pipelineState = m_luminancePipelineState.get();
                 drawItem.m_indexBufferView = &m_quadIndexBufferView;
-                drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+                drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
                 drawItem.m_shaderResourceGroups = shaderResourceGroups;
-                drawItem.m_streamBufferViewCount = static_cast<uint32_t>(m_quadStreamBufferViews[LuminanceMapScope].size());
+                drawItem.m_streamBufferViewCount = static_cast<uint8_t>(m_quadStreamBufferViews[LuminanceMapScope].size());
                 drawItem.m_streamBufferViews = m_quadStreamBufferViews[LuminanceMapScope].data();
                 commandList->Submit(drawItem);
             }
